Create a pull request template
Created a pull request template to:
Standardize and better manage the changes we make to our repo
Give others context of changes, and help others test any changes your code makes to the repo. This will go a long way to ensuring that we rigorously test our code to avoid introducing bugs in the system.